Fix a bug with rendering "/myBooks" on page Reload
When the component is rendered for the first time it loads only the books of current user. But on browser reload, there was a bug and loads all books from Database. The reason was that the database request happen in useEffect, and the collection of books was filtered in useEffect. After browser reload, useEffect do not run. The options was to get all books collection in useEffect and filtered it outside the useEffect.
The option that i chose is to put "user" in dependency array of the useEffect Hook.